refactor: adjust directory structure
