+++ packages/core/src/node/utils/index.ts
@@ -2,3 +2,7 @@ export * from './normalizePath';
 export * from './getPageKey';
 export * from './createHash';
 export * from './detectReactVersion';
+export * from './getASTNodeImport';
+export * from './flattenMdxContent';
+export * from './loadFrontMatter';
+export * from './applyReplaceRules';
