+++ packages/plugin-auto-nav-sidebar/src/index.ts
@@ -1,17 +1,7 @@
 import path from 'path';
-import fs from '@modern-js/utils/fs-extra';
-import {
-  NavItem,
-  RspressPlugin,
-  Sidebar,
-  SidebarGroup,
-  SidebarItem,
-  addTrailingSlash,
-  slash,
-  withBase,
-} from '@rspress/shared';
-import { logger } from '@rspress/shared/logger';
-import { NavMeta, SideMeta } from './type';
+import { RspressPlugin, addTrailingSlash } from '@rspress/shared';
+import { combineWalkResult } from './utils';
+import { walk } from './walk';
 
 // Scan all the directories and files in the work directory(such as `docs`), and then generate the nav and sidebar configuration according to the directory structure.
 // We will do as follows:
@@ -159,185 +149,6 @@ import { NavMeta, SideMeta } from './type';
 //   }
 // ]
 // ```
-export async function detectFilePath(rawPath: string) {
-  const extensions = ['.mdx', '.md', '.tsx', '.jsx', '.ts', '.js'];
-  // The params doesn't have extension name, so we need to try to find the file with the extension name.
-  let realPath: string | undefined = rawPath;
-  const filename = path.basename(rawPath);
-  if (filename.indexOf('.') === -1) {
-    const pathWithExtension = extensions.map(ext => `${rawPath}${ext}`);
-    const pathExistInfo = await Promise.all(
-      pathWithExtension.map(p => fs.pathExists(p)),
-    );
-    realPath = pathWithExtension.find((_, i) => pathExistInfo[i]);
-  }
-
-  if (!realPath) {
-    // Throw an error and it will be caught by the `extractH1Title`.
-    throw new Error();
-  }
-
-  return realPath;
-}
-
-export async function extractH1Title(filePath: string): Promise<string> {
-  try {
-    const realPath = await detectFilePath(filePath);
-    const content = await fs.readFile(realPath, 'utf-8');
-    const h1RegExp = /^#\s+(.*)$/m;
-    const match = content.match(h1RegExp);
-    return match ? match[1] : '';
-  } catch (e) {
-    logger.warn(
-      `Can't find the file: ${filePath}, please check it in "${path.join(
-        path.dirname(filePath),
-        '_meta.json',
-      )}".`,
-    );
-    return '';
-  }
-}
-
-export async function scanSideMeta(
-  workDir: string,
-  rootDir: string,
-  routePrefix: string,
-) {
-  const addRoutePrefix = (link: string) => `${routePrefix}${link}`;
-  // find the `_meta.json` file
-  const metaFile = path.resolve(workDir, '_meta.json');
-  // Fix the windows path
-  const relativePath = slash(path.relative(rootDir, workDir));
-  let sideMeta: SideMeta | undefined;
-  // Get the sidebar config from the `_meta.json` file
-  try {
-    // Don't use require to avoid require cache, which make hmr not work.
-    sideMeta = (await fs.readJSON(metaFile, 'utf8')) as SideMeta;
-  } catch (e) {
-    // If the `_meta.json` file doesn't exist, we will generate the sidebar config from the directory structure.
-    const subItems = await fs.readdir(workDir);
-    sideMeta = await Promise.all(
-      subItems.map(async item => {
-        const stat = await fs.stat(path.join(workDir, item));
-        // If the item is a directory, we will transform it to a object with `type` and `name` property.
-        if (stat.isDirectory()) {
-          return {
-            type: 'dir',
-            name: item,
-            label: item,
-          };
-        } else {
-          return item;
-        }
-      }),
-    );
-  }
-
-  const sidebarFromMeta: (SidebarGroup | SidebarItem)[] = await Promise.all(
-    sideMeta.map(async metaItem => {
-      if (typeof metaItem === 'string') {
-        const title = await extractH1Title(path.resolve(workDir, metaItem));
-        return {
-          text: title,
-          link: addRoutePrefix(
-            `${relativePath}/${metaItem.replace(/\.mdx?$/, '')}`,
-          ),
-        };
-      }
-
-      const {
-        type = 'file',
-        name,
-        label,
-        collapsible,
-        collapsed,
-        link,
-        tag,
-      } = metaItem;
-      if (type === 'file') {
-        const title =
-          label ?? (await extractH1Title(path.resolve(workDir, name)));
-        return {
-          text: title,
-          link: addRoutePrefix(
-            `${relativePath}/${name.replace(/\.mdx?$/, '')}`,
-          ),
-          tag,
-        };
-      } else if (type === 'dir') {
-        const subDir = path.resolve(workDir, name);
-        const subSidebar = await scanSideMeta(subDir, rootDir, routePrefix);
-        let realPath = '';
-        try {
-          realPath = await detectFilePath(subDir);
-        } catch (e) {
-          // ignore
-        }
-        return {
-          text: label!,
-          collapsible,
-          collapsed,
-          items: subSidebar,
-          link: realPath
-            ? addRoutePrefix(`${relativePath}/${name}`)
-            : undefined,
-          tag,
-        };
-      } else {
-        return {
-          text: label!,
-          link,
-          tag,
-        } as SidebarItem;
-      }
-    }),
-  );
-
-  return sidebarFromMeta;
-}
-
-export async function walk(workDir: string, routePrefix = '/') {
-  // find the `_meta.json` file
-  const rootMetaFile = path.resolve(workDir, '_meta.json');
-  let navConfig: NavMeta | undefined;
-  // Get the nav config from the `_meta.json` file
-  try {
-    navConfig = (await fs.readJSON(rootMetaFile, 'utf8')) as NavItem[];
-  } catch (e) {
-    navConfig = [];
-  }
-
-  navConfig.forEach(navItem => {
-    if ('items' in navItem) {
-      navItem.items.forEach(item => {
-        if ('link' in item) {
-          item.link = withBase(item.link, routePrefix);
-        }
-      });
-    }
-    if ('link' in navItem) {
-      navItem.link = withBase(navItem.link, routePrefix);
-    }
-  });
-  // find the `_meta.json` file in the subdirectory
-  const subDirs = (await fs.readdir(workDir)).filter(v =>
-    fs.statSync(path.join(workDir, v)).isDirectory(),
-  );
-  // Every sub dir will represent a group of sidebar
-  const sidebarConfig: Sidebar = {};
-  for (const subDir of subDirs) {
-    const sidebarGroupKey = `${routePrefix}${subDir}`;
-    sidebarConfig[sidebarGroupKey] = await scanSideMeta(
-      path.join(workDir, subDir),
-      workDir,
-      routePrefix,
-    );
-  }
-  return {
-    nav: navConfig as NavItem[],
-    sidebar: sidebarConfig,
-  };
-}
 
 function processLocales(
   langs: string[],
@@ -365,19 +176,7 @@ function processLocales(
               addTrailingSlash(lang === defaultLang ? '' : `/${lang}`),
             ),
           ];
-      return walks.reduce(
-        (acc, cur, curIndex) => ({
-          nav: {
-            ...acc.nav,
-            [versions[curIndex] || 'default']: cur.nav,
-          },
-          sidebar: { ...acc.sidebar, ...cur.sidebar },
-        }),
-        {
-          nav: {},
-          sidebar: {},
-        },
-      );
+      return combineWalkResult(walks, versions);
     }),
   );
 }
@@ -420,19 +219,7 @@ export function pluginAutoNavSidebar(): RspressPlugin {
             )
           : [await walk(config.root!)];
 
-        const combined = walks.reduce(
-          (acc, cur, curIndex) => ({
-            nav: {
-              ...acc.nav,
-              [versions[curIndex] || 'default']: cur.nav,
-            },
-            sidebar: { ...acc.sidebar, ...cur.sidebar },
-          }),
-          {
-            nav: {},
-            sidebar: {},
-          },
-        );
+        const combined = combineWalkResult(walks, versions);
 
         config.themeConfig = { ...config.themeConfig, ...combined };
       }
